New file for cavern door funcs on psx
